PanTilt working.  Movement Advanced added, Bluetooth Control app added.

+++ TinkerTank/Movement/IMovementInterface.cs
@@ -12,7 +12,7 @@ namespace Peripherals
         DriveMethod driveMethod { get; }
         ComponentStatus Init(Meadow.Hardware.IPin HBridge1PinA, Meadow.Hardware.IPin HBridge1Pinb, Meadow.Hardware.IPin HBridge1PinEnable,
             Meadow.Hardware.IPin HBridge2PinA, Meadow.Hardware.IPin HBridge2Pinb, Meadow.Hardware.IPin HBridge2PinEnable);
-        void Move(Direction direction, float power, TimeSpan? movementDuration = null);
+        void Move(Direction direction, int power, TimeSpan? movementDuration = null);
         bool MoveManual(float leftPower, float rightPower, TimeSpan? movementDuration = null);
         bool MoveManual(float leftFrontPower, float rightFrontPower, float leftRearPower, float rightRearPower, TimeSpan? movementDuration = null);
         void SetDefaultPower(int defaultPower);
